refactor: split large lock repair helpers
Move the large-lock repair-issue and acknowledgement-store logic to a dedicated
large_lock_repairs.py module so helpers.py drops below the file-size threshold.

Rename _supports_connection_status to public provider_supports_connection_status
because repairs.py uses it across modules. Behavior is otherwise unchanged.

Closes #772
